+++ jidokhae-web/src/app/(admin)/admin/members/page.tsx
@@ -1,11 +1,12 @@
 import { redirect } from 'next/navigation'
 import { createClient } from '@/lib/supabase/server'
+import { getUser } from '@/lib/auth'
 import { getProfile } from '@/lib/profile'
 import MemberList from '@/components/admin/MemberList'
 
 export default async function MembersPage() {
   const supabase = await createClient()
-  const { data: { user } } = await supabase.auth.getUser()
+  const user = await getUser()
   if (!user) redirect('/auth/login')
 
   const profile = await getProfile(user.id)
